requirements.txt
flask
flask
yfinance
google-generativeai
pymongo
matplotlib
streamlit
plotly
pandas
python-dotenv